// src/shared/pages/task/type.ts
export interface RowsItem {
  psubject: string
  pstatus: string
  fcn: string
  pup: string
  pend_date: string
  preg_user: string
  taseq: string
  cn: string
  type: string
  subject: string
  ticket_refer: string
  status: string
  priority: string
  parent_task: string
  start_date: string
  due_date: string
  assignee: null
  acn: string
  percent: string
  estimated_time: string
  time_zone: string
  userno: string
  ucn: string
  files: string
  options: null
  mime: string
  mcn: string
  mchecksum: string
  reg_date: string
  pcn: string
  pseq: string
  sub_task: null
  hashtags: null
  relate_task: null
  reference_doc: null
  updated_date: string
  country_zone: string
  dept: string
  name: string
  rank: string
  sdept: string
  sname: string
  srank: string
  fdept: string
  fname: string
  frank: string
  odept: null
  oname: null
  orank: null
  osdept: null
  osname: null
  osrank: null
  ofdept: null
  ofname: null
  ofrank: null
  ndept: null
  nname: null
  nrank: null
  nsdept: null
  nsname: null
  nsrank: null
  nfdept: null
  nfname: null
  nfrank: null
  country: string
  dcountry: string
  ocountry: string
  odcountry: string
  ncountry: string
  ndcountry: string
  is_read: string
  notify_type: string
  sseq: null
  scn: null
  releases: null
  story_point: string
  epic_label_cn: null
  epic_label_id: null
  value_point: string
  constraint_condition: null
  done_date: null
  release: string
  taseq_no: string
  seq_no: string
  tab_id: string
  un_read: boolean
  no: number
  type_msg: string
  status_msg: string
  priority_msg: string
}

export interface Page {
  total_num: string
  total_page: number
}

export interface Root {
  success: boolean
  rows: Array<RowsItem>
  page: Page
  benchmark: Array<string>
}